requirements.txt
requests
edge-tts
openai
ffmpeg-python
python-dotenv
asyncio
moviepy
instagrapi
